chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-style-basic-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE ERROR: line 42: Uncaught TypeError: Cannot read property 'createElement' of undefined
PASS document.styleSheets.length is 2
FAIL importLink.import.styleSheets.length should be 1. Threw exception TypeError: Cannot read property 'styleSheets' of undefined
FAIL window.getComputedStyle(shouldBeBlue).color should be rgb(0, 0, 255). Was rgb(0, 0, 0).
FAIL window.getComputedStyle(shouldBeYellow).color should be rgb(255, 255, 0). Was rgb(0, 0, 0).
PASS window.getComputedStyle(shouldBeWhite).color is "rgb(255, 255, 255)"
FAIL window.getComputedStyle(shouldBeTeal).color should be rgb(0, 128, 128). Was rgb(0, 0, 0).
PASS window.getComputedStyle(shouldBeGreen).color is "rgb(0, 0, 0)"
FAIL Unexpected error: Uncaught TypeError: Cannot read property 'createElement' of undefined
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

These elements should be styled appropriately:
Should be blue.
Should be white.
Should be yellow.
Should be teal.
Should become green.
Should become aqua.
